+++ TrainingExtensions/torch/test/python/experimental/v2/models_/models_to_test.py
@@ -39,6 +39,9 @@
 import torch
 from torch import nn
 
+from aimet_torch import elementwise_ops
+
+
 class SimpleConditional(torch.nn.Module):
     """
     Model using conditional paths
@@ -144,3 +147,251 @@ class SoftMaxAvgPoolModel(torch.nn.Module):
     def forward(self, inp):
         x = self.sfmax(inp)
         return self.avgpool(x)
+
+
+class SingleResidual(nn.Module):
+    """ A model with a single residual connection.
+        Use this model for unit testing purposes. """
+
+    def __init__(self, num_classes=10):
+        super(SingleResidual, self).__init__()
+        self.conv1 = nn.Conv2d(3, 32, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn1 = nn.BatchNorm2d(32)
+        self.relu1 = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=2, stride=2, padding=1)
+        # All layers above are same as ResNet
+        # The output of the MaxPool2d is used as a residual.
+
+        # The following layers are considered as single block.
+        self.conv2 = nn.Conv2d(32, 16, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn2 = nn.BatchNorm2d(16)
+        self.relu2 = nn.ReLU(inplace=True)
+        self.conv3 = nn.Conv2d(16, 8, kernel_size=2, stride=2, padding=2, bias=False)
+
+        # The output of Conv2d layer above(conv3) is added with the the residual from
+        # MaxPool2d and then fed to the relu layer below.
+        self.relu3 = nn.ReLU(inplace=True)
+
+        self.avgpool = nn.AvgPool2d(3, stride=1)
+        self.conv4 = nn.Conv2d(32, 8, kernel_size=2, stride=2, padding=2, bias=True)
+        self.ada = nn.AdaptiveAvgPool2d(5)
+        self.fc = nn.Linear(72, num_classes)
+
+    def forward(self, *inputs):
+        x = self.conv1(inputs[0])
+        x = self.bn1(x)
+        x = self.relu1(x)
+        x = self.maxpool(x)
+
+        # Save the output of MaxPool as residual.
+        residual = x
+
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu2(x)
+        x = self.conv3(x)
+
+        # Add the residual
+        # AdaptiveAvgPool2d is used to get the desired dimension before adding.
+        residual = self.conv4(residual)
+        residual = self.ada(residual)
+        x += residual
+        x = self.relu3(x)
+
+        x = self.avgpool(x)
+        x = x.view(x.size(0), -1)
+        x = self.fc(x)
+        return x
+
+
+class SingleResidualWithAvgPool(nn.Module):
+    """ A model with a single residual connection.
+        Use this model for unit testing purposes. """
+
+    def __init__(self, num_classes=10):
+        super(SingleResidualWithAvgPool, self).__init__()
+        self.conv1 = nn.Conv2d(3, 32, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn1 = nn.BatchNorm2d(32)
+        self.relu1 = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=2, stride=2, padding=1)
+        # All layers above are same as ResNet
+        # The output of the MaxPool2d is used as a residual.
+
+        # The following layers are considered as single block.
+        self.conv2 = nn.Conv2d(32, 16, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn2 = nn.BatchNorm2d(16)
+        self.relu2 = nn.ReLU(inplace=True)
+        self.conv3 = nn.Conv2d(16, 8, kernel_size=2, stride=2, padding=2, bias=False)
+
+        # The output of Conv2d layer above(conv3) is added with the the residual from
+        # MaxPool2d and then fed to the relu layer below.
+        self.relu3 = nn.ReLU(inplace=True)
+
+        self.avgpool = nn.AvgPool2d(3, stride=1)
+        self.conv4 = nn.Conv2d(32, 8, kernel_size=2, stride=2, padding=2, bias=True)
+        self.ada = nn.AvgPool2d(5)
+        self.fc = nn.Linear(72, num_classes)
+
+    def forward(self, *inputs):
+        x = self.conv1(inputs[0])
+        x = self.bn1(x)
+        x = self.relu1(x)
+        x = self.maxpool(x)
+
+        # Save the output of MaxPool as residual.
+        residual = x
+
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu2(x)
+        x = self.conv3(x)
+
+        # Add the residual
+        # AdaptiveAvgPool2d is used to get the desired dimension before adding.
+        residual = self.conv4(residual)
+        residual = self.ada(residual)
+        x += residual
+        x = self.relu3(x)
+
+        x = self.avgpool(x)
+        x = x.view(x.size(0), -1)
+        x = self.fc(x)
+        return x
+
+
+class SingleResidualWithModuleAdd(nn.Module):
+    """ A model with a single residual connection.
+        Use this model for unit testing purposes. """
+
+    def __init__(self, num_classes=10):
+        super(SingleResidualWithModuleAdd, self).__init__()
+        self.conv1 = nn.Conv2d(3, 32, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn1 = nn.BatchNorm2d(32)
+        self.relu1 = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=2, stride=2, padding=1)
+        # All layers above are same as ResNet
+        # The output of the MaxPool2d is used as a residual.
+
+        # The following layers are considered as single block.
+        self.conv2 = nn.Conv2d(32, 16, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn2 = nn.BatchNorm2d(16)
+        self.relu2 = nn.ReLU(inplace=True)
+        self.conv3 = nn.Conv2d(16, 8, kernel_size=2, stride=2, padding=2, bias=False)
+
+        # The output of Conv2d layer above(conv3) is added with the the residual from
+        # MaxPool2d and then fed to the relu layer below.
+        self.add = elementwise_ops.Add()
+        self.relu3 = nn.ReLU(inplace=True)
+
+        self.avgpool = nn.AvgPool2d(3, stride=1)
+        self.conv4 = nn.Conv2d(32, 8, kernel_size=2, stride=2, padding=2, bias=True)
+        self.ada = nn.AdaptiveAvgPool2d(5)
+        self.fc = nn.Linear(72, num_classes)
+
+    def forward(self, *inputs):
+        x = self.conv1(inputs[0])
+        x = self.bn1(x)
+        x = self.relu1(x)
+        x = self.maxpool(x)
+
+        # Save the output of MaxPool as residual.
+        residual = x
+
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu2(x)
+        x = self.conv3(x)
+
+        # Add the residual
+        # AdaptiveAvgPool2d is used to get the desired dimension before adding.
+        residual = self.conv4(residual)
+        residual = self.ada(residual)
+        x = self.add(x, residual)
+        x = self.relu3(x)
+
+        x = self.avgpool(x)
+        x = x.view(x.size(0), -1)
+        x = self.fc(x)
+        return x
+
+
+class MultiInput(nn.Module):
+    """ A model with multiple inputs.
+        Use this model for unit testing purposes. """
+
+    def __init__(self, num_classes=3):
+        super(MultiInput, self).__init__()
+        self.conv1 = nn.Conv2d(3, 16, kernel_size=2, stride=2, padding=3, bias=False)
+        self.conv2 = nn.Conv2d(16, 8, kernel_size=3, stride=2, padding=2)
+        self.conv3 = nn.Conv2d(3, 8, kernel_size=3, stride=2, padding=2)
+        self.bn1 = nn.BatchNorm2d(8)
+        self.relu1 = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=2, stride=2, padding=1)
+        self.fc = nn.Linear(288, num_classes)
+
+    def forward(self, *inputs):
+        x1 = self.conv1(inputs[0])
+        x1 = self.conv2(x1)
+        x2 = self.conv3(inputs[1])
+        x = x1 + x2
+        x = self.bn1(x)
+        x = self.relu1(x)
+        x = self.maxpool(x)
+        x = x.view(x.size(0), -1)
+        x = self.fc(x)
+        return x
+
+
+class ModelWithBertCustomLayerNormGelu(torch.nn.Module):
+    """ Model with PyTorch LayerNorm and gelu """
+
+    def __init__(self):
+        super().__init__()
+        self.linear1 = torch.nn.Linear(4, 4)
+        # default attribute -
+        # eps = 1e-05 and elementwise_affine = True
+        # parameters : weight and bias
+        self.customln1 = torch.nn.LayerNorm(4)
+        self.gelu1 = torch.nn.GELU()
+
+    def forward(self, x):
+        x = self.linear1(x)
+        x = self.customln1(x)
+        x = self.gelu1(x)
+        return x
+
+
+class QuantSimTinyModel(nn.Module):
+    """ Use this model for quantsim_config unit testing purposes. Expect input shape (1, 3, 32, 32) """
+
+    def __init__(self):
+        super(QuantSimTinyModel, self).__init__()
+        self.conv1 = nn.Conv2d(3, 32, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn1 = nn.BatchNorm2d(32)
+        self.relu1 = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=2, stride=2, padding=1)
+        self.conv2 = nn.Conv2d(32, 16, kernel_size=2, stride=2, padding=2, bias=False)
+        self.bn2 = nn.BatchNorm2d(16)
+        self.relu2 = nn.ReLU6(inplace=True)
+        self.conv3 = nn.Conv2d(16, 8, kernel_size=2, stride=2, padding=2, bias=False)
+        self.relu3 = nn.ReLU(inplace=True)
+        self.avgpool = nn.AvgPool2d(3, stride=1)
+        self.conv4 = nn.Conv2d(8, 4, kernel_size=2, stride=2, padding=2, bias=True)
+        self.fc = nn.Linear(36, 12)
+
+    def forward(self, *inputs):
+        x = self.conv1(inputs[0])
+        x = self.bn1(x)
+        x = self.relu1(x)
+        x = self.maxpool(x)
+
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu2(x)
+        x = self.conv3(x)
+        x = self.relu3(x)
+        x = self.avgpool(x)
+        x = self.conv4(x)
+        x = x.view(x.size(0), -1)
+        x = self.fc(x)
+        return x
